feat: Refactor UI components and add new ones
- Updated Button component to use class-variance-authority for styling variants.
- Refactored Card component and its subcomponents (Header, Title, Description, Content, Footer) for improved structure and styling.
- Introduced Badge component with variant support.
- Added Calendar component with custom day button and styling.
- Created Checkbox component using Radix UI.
- Implemented Dialog component with overlay, content, header, footer, title, and description.
- Added Label component for form elements.
- Introduced Popover component for contextual overlays.
- Added Toaster component for notifications using Sonner.
- Created useToast hook for simplified toast notifications.
- Updated package.json to include new dependencies for Radix UI components, Sonner, and react-day-picker.

+++ backend/Mizan.Tests/Application/LogFoodCommandTests.cs
@@ -72,7 +72,7 @@ public class LogFoodCommandTests : IDisposable
         // Assert
         result.Should().NotBeNull();
         result.Id.Should().NotBeEmpty();
-        result.Calories.Should().Be((int)(165 * 1.5));
+        result.Calories.Should().Be(165 * 1.5m);
         result.ProteinGrams.Should().Be(31 * 1.5m);
         result.Message.Should().Contain("Chicken Breast");
